make templates configurable from a file
The templates are now configurable via a Lua file returning the
translation table. It is 'qtemplates.lua' by default, but a custom file
can be specified on the command line using the '-c filepath' switch.

The keys in the table can be either the names of template classes,
or the module name (qtcore, qtgui...). See generator/qtemplates.lua
for examples of usage.