First attempt at safe virtual overrides
Virtual methods implemented in Lua are not called
until a flag is set in the bitset hasOverride in
the shell class. This bitfield is updated in the
__newindex by checking if the value is a function
and key is a name known at compile time.

The overhead at __newindex is lower lower than
actually checking for overrides at runtime, and
also possibly safer - no Lua calls are made (apart
from lua_gettop) when the override is not present.